native/test: Reinstate nofib test suite; move fusion playground to test/fusion
