+++ accelerate-llvm-native/accelerate-llvm-native.cabal
@@ -216,6 +216,30 @@ test-suite nofib-llvm-native
         -with-rtsopts=-n4M
         -with-rtsopts=-N
 
+test-suite fusion-tests-llvm-native
+  type:                 exitcode-stdio-1.0
+  hs-source-dirs:       test/fusion
+  main-is:              Main.hs
+
+  build-depends:
+          base                          >= 4.10
+        , accelerate
+        , accelerate-llvm-native
+        , criterion
+        , deepseq
+
+  default-language:
+        Haskell2010
+
+  ghc-options:
+        -Wall
+        -O2
+        -threaded
+        -rtsopts
+        -with-rtsopts=-A128M
+        -with-rtsopts=-n4M
+        -with-rtsopts=-N
+
 
 source-repository head
   type:                 git
